Add zoom to map editor and speed to skill editor, decimal to speed

+++ src/Core/Objects/JobBase.cs
@@ -19,7 +19,7 @@ namespace Core.Objects
             FemaleSprite = 0;
             MaleSprite = 0;
             StartMap = 1;
-            MoveSpeed = 1;
+            MoveSpeed = 1.0f;
 
             for (int i = 0; i < Core.Globals.Variables.MaxStartItems; i++)
             {
@@ -45,7 +45,7 @@ namespace Core.Objects
         public byte StartX;
         public byte StartY;
         public int BaseExp;
-        public int MoveSpeed;
+        public float MoveSpeed;
 
         public static List<JobBase> Instance { get; private set; } = new List<JobBase>();
